third_party/blink/web_tests/media/mediasession/mojo/media-control-set-handler-notifies-service.html

<!DOCTYPE html>
<title>Test that setting MediaSession event handler should notify the service</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/services/media_session/public/mojom/media_session.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/mediasession/media_session.mojom.js"></script>
<script src="resources/mediasessionservice-mock.js"></script>
<script src="resources/utils.js"></script>
<script>

var expectations;

var nextExpectation = 0;

// Delay the initialization since the initialization of MediaSessionAction is delayed.
function getExpectations() {
  if (!expectations) {
    expectations = [
      [ mediaSession.mojom.MediaSessionAction.kPlay, true ],
      [ mediaSession.mojom.MediaSessionAction.kPause, true ],
      [ mediaSession.mojom.MediaSessionAction.kPreviousTrack, true ],
      [ mediaSession.mojom.MediaSessionAction.kNextTrack, true ],
      [ mediaSession.mojom.MediaSessionAction.kSeekBackward, true ],
      [ mediaSession.mojom.MediaSessionAction.kSeekForward, true ],
      [ mediaSession.mojom.MediaSessionAction.kSeekTo, true ],
      [ mediaSession.mojom.MediaSessionAction.kPlay, false ],
      [ mediaSession.mojom.MediaSessionAction.kPause, false ],
      [ mediaSession.mojom.MediaSessionAction.kPreviousTrack, false ],
      [ mediaSession.mojom.MediaSessionAction.kNextTrack, false ],
      [ mediaSession.mojom.MediaSessionAction.kSeekBackward, false ],
      [ mediaSession.mojom.MediaSessionAction.kSeekForward, false ],
      [ mediaSession.mojom.MediaSessionAction.kSeekTo, false ],
      [ mediaSession.mojom.MediaSessionAction.kPlay, true ],
      [ mediaSession.mojom.MediaSessionAction.kPause, true ],
      [ mediaSession.mojom.MediaSessionAction.kPreviousTrack, true ],
      [ mediaSession.mojom.MediaSessionAction.kNextTrack, true ],
      [ mediaSession.mojom.MediaSessionAction.kSeekBackward, true ],
      [ mediaSession.mojom.MediaSessionAction.kSeekForward, true ],
      [ mediaSession.mojom.MediaSessionAction.kSeekTo, true ],
    ];
  }
  return expectations;
}

async_test(function(t) {
  let m = mediaSessionServiceMock;
  m.setEnableDisableActionCallback(t.step_func(function(action, isEnabled) {
    var expectedAction = getExpectations()[nextExpectation][0];
    var expectedIsEnabled = getExpectations()[nextExpectation][1];
    assert_equals(expectedAction, action);
    assert_equals(expectedIsEnabled, isEnabled);
    if (++nextExpectation >= getExpectations().length)
      t.done();
  }));

  // Setting handlers should notify the service.
  window.navigator.mediaSession.setActionHandler("play", _ => {});
  window.navigator.mediaSession.setActionHandler("pause", _ => {});
  window.navigator.mediaSession.setActionHandler("previoustrack", _ => {});
  window.navigator.mediaSession.setActionHandler("nexttrack", _ => {});
  window.navigator.mediaSession.setActionHandler("seekbackward", _ => {});
  window.navigator.mediaSession.setActionHandler("seekforward", _ => {});
  window.navigator.mediaSession.setActionHandler("seekto", _ => {});

  // Setting handlers again should not notify the service.
  window.navigator.mediaSession.setActionHandler("play", _ => {});
  window.navigator.mediaSession.setActionHandler("pause", _ => {});
  window.navigator.mediaSession.setActionHandler("previoustrack", _ => {});
  window.navigator.mediaSession.setActionHandler("nexttrack", _ => {});
  window.navigator.mediaSession.setActionHandler("seekbackward", _ => {});
  window.navigator.mediaSession.setActionHandler("seekforward", _ => {});
  window.navigator.mediaSession.setActionHandler("seekto", _ => {});

  // Unsetting handlers should notify the service.
  window.navigator.mediaSession.setActionHandler("play", null);
  window.navigator.mediaSession.setActionHandler("pause", null);
  window.navigator.mediaSession.setActionHandler("previoustrack", null);
  window.navigator.mediaSession.setActionHandler("nexttrack", null);
  window.navigator.mediaSession.setActionHandler("seekbackward", null);
  window.navigator.mediaSession.setActionHandler("seekforward", null);
  window.navigator.mediaSession.setActionHandler("seekto", null);

  // Unsetting handlers should not notify the service.
  window.navigator.mediaSession.setActionHandler("play", null);
  window.navigator.mediaSession.setActionHandler("pause", null);
  window.navigator.mediaSession.setActionHandler("previoustrack", null);
  window.navigator.mediaSession.setActionHandler("nexttrack", null);
  window.navigator.mediaSession.setActionHandler("seekbackward", null);
  window.navigator.mediaSession.setActionHandler("seekforward", null);
  window.navigator.mediaSession.setActionHandler("seekto", null);

  // Setting handlers again should notify the service.
  window.navigator.mediaSession.setActionHandler("play", _ => {});
  window.navigator.mediaSession.setActionHandler("pause", _ => {});
  window.navigator.mediaSession.setActionHandler("previoustrack", _ => {});
  window.navigator.mediaSession.setActionHandler("nexttrack", _ => {});
  window.navigator.mediaSession.setActionHandler("seekbackward", _ => {});
  window.navigator.mediaSession.setActionHandler("seekforward", _ => {});
  window.navigator.mediaSession.setActionHandler("seekto", _ => {});
}, "test that setting event handler notifies the mojo service");

</script>
